[POI-816]RPW]Improving Global Settings lifetime
If a `IRenderPipelineGraphicsSettings `is on the settings list from a `RenderPipelineGlobalSettings`, and the user edits the .asset outside the Unity editor, the `settingsList `are not auto populated, leading to errors if the pipeline expect to have some settings to be present on the list.

As the settings have changed externally, we are recreating the pipeline, in order that the pipeline obtains the default values back again from the `GraphicsSettings`.

https://jira.unity3d.com/browse/XPIPELINE-799

Besides that I've covered multiple areas that were uncovered.

1.Delete from Unity will show a warning to the user:

![image](https://media.github.cds.internal.unity3d.com/user/3279/files/8f1e842a-b54b-452f-9246-07c56fd5a1eb)

2. Deleting from outside Unity will recreate the pipeline.
3. GraphicsSettings outside modifications was only ensuring global settings instead of recreating  the pipeline, this might lead to errors as the referenced asset could have potentially changed.
4.  When changing the global settings for the current pipeline, the pipeline is recreated, this is handled for UI and for code.

As we are ensuring the pipeline recreation when is need, we do not need to check and potentially ensure the global settings on each render of the pipeline.

## QA Needed tests
1.Check deleteing the global settings from outside and inside unity
2.Check URP and HDRP templates
3.Create a project from scratch and then add the HDRP package, and set up hdrp, see that the global settings are created.
4, Do step 3 but with URP.
5.Move global settings assets,
6.Go to graphics settings outside unity, remove the references of the m_RenderPipelienGlobalSettings. See that the asset pipeline is recreated.