[VFX] Template window multiple fixes
### UUM-47192
When there's no user defined category the built-in category cannot be collapsed so that the window cannot feel "empty" anymore.

### UUM-47453
When moving the splitter to reduce the width of the left panel we could reach a point where the width of each item is uneven. Now the left panel has a minimum size of 200px and items should all have the same width, always.

### UUM-47505
The template item hit box was slightly bigger than its visual representation, now it perfectly match.

### UUM-47506
The right panel (details panel) could be resize to as small as zero width, leading to messy layout. Now it has a minimum width of 200px (like the left panel).