Fix Directional light PCSS API
Changing directional light API to make maximum blur size consistent for different angular diameters

Replacing "Max Blocker Distance" parameter with "Max Penumbra Size".  
The maximum blocker distance is now computed from the max penumbra size based on the light angular diameter, so the sampling is constrained similarly when angular diameter is changed.  
The default max blocker distance of 64m was tuned based on a light diameter of 0.5 degrees.  When using a larger light diameter (such as in the HDRP default template where it is set to 2.5 degrees), this made the filter kernel get way larger at that distance which reduces shadow quality and introduce artifacts close to the caster as seen in these before/after shots:
![Before](https://media.github.cds.internal.unity3d.com/user/1727/files/b8821b55-a625-4fc3-91d5-cdc213d9d9b7)
![After](https://media.github.cds.internal.unity3d.com/user/1727/files/40b04edf-b9b8-4f1c-a096-5d7066b5d548)

Also exposing softness parameter that was missing in the UI for directional lights even though the value is used