Bugfix cascaded shadow maps popping on first cascade when shadowmask is on
Bugfix cascaded shadow maps popping on first cascade when shadowmask is on.
This regression was introduced by  https://github.cds.internal.unity3d.com/unity/unity/pull/35432

See the gif below that shows the bug.
![bug shadow popping](https://media.github.cds.internal.unity3d.com/user/1208/files/4342aba4-b3e4-4739-ba81-38303cd5612a)

I modified the HDRP test 2309 (Shadowmask) so it covers the transition with last cascade and shadowmask as well as the transition with cascade 0 and cascade 1.
This is the cascade coverage:
![after-cascades](https://media.github.cds.internal.unity3d.com/user/1208/files/42a214ce-d8d7-4b66-a361-8af5a4fac9b9)

This is before the fix:
![before](https://media.github.cds.internal.unity3d.com/user/1208/files/6c87a72d-c0dc-479a-ad63-c3eafbff4620)

After the fix

![after](https://media.github.cds.internal.unity3d.com/user/1208/files/64bbcc87-2cbd-440f-b556-f5f99ae93b88)