[VFX] Allow instanced rendering of VFX that have exposed textures or graphic buffers
Second half of the feature that splits batches when there are textures or buffers that may be different between instances: https://jira.unity3d.com/browse/VFXG-366
First half PR here: https://github.cds.internal.unity3d.com/unity/unity/pull/23677
This part takes care of optimizing the rendering in this circumstances. With the previous PR, when this happened, every instance was rendered individually, which means that it had to set up the shader and all the uniform properties.
With these changes, now they are still considered a single batch by the instancing batcher, setting up the shader just once, and it is responsibility of the VFX render command to prepare the bindings for the textures or graphic objects that may change for a specific drawcall.
To do this, it was necessary to refactor the VFXRenderData system, to allow allocating more than one per VFX command.
To validate the improvement, it is necessary to compare render captures of a scene that contains many VFX with some different values for a exposed texture.
Testing the performance of such scene with and without the changes should also show some difference.