Graphics/SRP/RPC  Extended UI Overlay Support with SRPs
_Currently screen space overlay rendering support is limited in Scriptable Render Pipelines (SRPs). It has been partially implemented to support High Dynamic Range (HDR) in SRPs, but is not enabled for non-HDR cases, and only focuses on Unity UI (uGUI) and UI Toolkit overlays. Coverage of other overlay features (watermark, IMGUI, development console...) from SRPs is missing, they are rendered by native side on top of SRPs output. This lack of control is producing bugs, notably with HDR, or performance issues by preventing some potential optimizations as SRPs are not even aware of the overlays existence._

Current PR focuses on:
1. Extending UI Overlay support by bringing IMGUI and Dev console under SRPs control, SRPs already have uGUI (Canvas) and UI Toolkit coverage.
2. Enabling by default UI Overlay rendering from URP and HDRP (and not from native side after SRP rendering), for both HDR and non-HDR modes (without including XR and URP NRP, see comment below).

Following discussions with XR SRP and HDR Sony/Nintendo platforms teams, screen space UI Overlay will still be triggered by native side for XR cases. 

For URP with Native Render Pass (NRP) without Render Graph, we force pure native rendering as NRP doesn't support setRenderTarget() used in URP Screen Space UI Pass.

Note that the changes are focusing on SRP and NOT on IMGUI which is currently a KTOL (Keep The Light On) area. We are mostly aiming at improving SRP rendering coverage in the render loop when a SRP is selected. We do it for multiple reasons:
 1. Parity with BiRP
 2. SRP rendering optimization with potential merging of sub render passes through Render Graph
 3. Various bug fixes.