[Port] [2023.2] [HDRP] Various fixes and cleanup of rectangular area lights code
The rectangular area light implementation of the dual-lobe SSS Lit material has been fixed as it was incorrect (Jira 47070) and cookie area light support have been added to water.

This PR also factor the rectangular area light code as it is a source of various mistake. Here is a list of change:
* PolygonFormFactor() now supports 3-5 vertices. Duplicate implementations have been removed.
* PolygonIrradiance() now always outputs the vector irradiance. Duplicate implementations have been removed. Warning: depending on whether the sphere light approximation is enabled, the vector irradiance may correspond either to the horizon-clipped or to the original polygon. This must be addressed in the future.
* The variant of SampleAreaLightCookie() without the roughness parameter has been removed. It is deemed too dangerous, since there are multiple instances of developers forgetting to pass the correct value.
* IntegrateEdge() has been removed.
* EvaluateLTC_Rect() has been added, which significantly reduces the effort (typing) required to support the rectangular area lights.
* The rectangular area light implementation of the dual-lobe SSS Lit material has been fixed.
* Most materials were not passing the correct roughness value to the function responsible for evaluation of the rectangular area light "cookie". This has been fixed.
* Rectangular area light "cookie" support has been added to the water shader.
* The AxF material now shares the implementation of the rectangular area lights with the other materials.

In addition, it is impossible to disable the sphere light approximation for the rectangular area lights - the code simply does not compile. Doing so is very useful for debugging, since this approximation often causes issues.This PR fixed it.

In addition, the details of the implementation of the rectangular area lights "leak" into the material files (e.g. look for APPROXIMATE_POLY_LIGHT_AS_SPHERE_LIGHT in Lit.hlsl).

The new code is shorter, simpler, and more robust.