Graphics/hd/reflection probe importance
Technical and Lighting Artists expressed a need to have an importance setting in HDRP reflection probes to better sort them.
URP and BiRP reflection probes both have an importance setting that allows to sort them, probes of higher importance being displayed "on top" of probes with lower importance.
HDRP in comparison is sorting solely based on the probe volume, where probes with a lower volume are displayed on top.

This PR implements an importance setting that acts the same way as in URP and BiRP.

Details:
Probes sorting was based on a uint sortKey value storing multiple information of the probe, where the volume value was stored in the 20 bits of higher weight.
It was not possible to add the importance value to those 20 bits without greatly reducing the maximum stored volume value or reducing the precision.
So it was decided to change the sortKey to ulong. Resulting in 52 bits available to store the volume/importance, the following changes were made:
 - Change the way of storing the volume to directly use the volume value instead of using a logarithmic scale. Store in 31 bits instead of 20 (1st sign bit is ignored as volume is always positive)
 - The next 15 bits of higher weight store the importance. This matches the 0 to 32767 range for importance used in the other pipelines. The value is stored in reverse order (32767 - importance) for proper ordering with volume size.
 - This still leaves 6 bits for further usage if necessary.

The default importance value is 1 (like in the other pipelines), allowing the easilly reduce the priority of a probe's display and send it "bellow" other ones if necessary.

Probes of the same importance are still ordered by volume between them, resulting in no rendering change when updating a project from an earlier version.

Planar reflection probes are sorted the same way, but: As in previous behaviour, planar probes were created by default with a very small Y value in the volume, they automatically resulted in a very high priority. It is now possible to have a cube reflection probe rendered over a planar reflection probe by using the importance value.
However, as planar reflection probes are more accurate in most of the cases, they have an importance value of 64 by default, allowing to filter many (or hopefully way enough) cube probes before overriding a planar probe.

Test 2201 was updated to also cover importance ordering : 

| before | after |
| ------- | ----- |
| ![PR36107_01](https://media.github.cds.internal.unity3d.com/user/696/files/f510c575-12a2-4452-a2ec-2d1def9bc65d) | ![PR36107_02](https://media.github.cds.internal.unity3d.com/user/696/files/fa2c2995-cea3-4bb8-b6f4-4c8cdcdcac64) |