Use the statically-linked runtime library for the native denoiser DLL
https://jira.unity3d.com/projects/UNDB/issues/UNDB-1
We now use a multi-threaded statically-linked runtime library (-MT or equivalent flag) for the native plugin binaries of the denoising package. The functionality has not changed, just the way we do the linking.

Note: the native binaries are built in the denoising package repo: https://github.com/Unity-Technologies/UnityDenoisingPlugin/pull/11