[VFX] [Strips] Indirect draw and sorting
https://jira.unity3d.com/browse/VFXG-78
Indirect draw mode is an optimization that allows to render only the active particles by collecting their indices on a list during update, and using that list as an indirection during rendering.
Before this PR, strips are rendering strip capacity * (particle per strip - 1) particles. Vertices for inactive particles will be discarded. This is acceptable when the system is using all or most of the particles, but it could be inefficient if the strip count or particle count is variable.
This PR is reusing the existing code for particles, adapting it to work with strips as well.
One thing to notice is that the last particle of each strip will be rendered and discarded, as we have one particle more than strip segments. This plus the extra indirection make this mode less efficient when we are using full capacity.

Using indirect draw to render particles in a different order also enables per-particle sorting. Sorting works for strips the same as for particle systems, and shares most of the code. Strips are built using 2 particles, the first one is used for sorting.