Renderer Feature Filtering
Solves the issue of Renderer Features showing up as useable for renderers they do not work with by adding an attribute to the Renderer Feature class. This attribute specifies which renderer(s) the Renderer Feature can work with. Not using this attribute on a Renderer Feature will allow it to continue to work as before.

https://jira.unity3d.com/browse/D2D-5315