Update LOD_FADE_CROSSFADE to affect all stages in URP
This PR updates the LOD_FADE_CROSSFADE keyword to compile for all stages in URP handwritten shaders.

This change was already applied to shadergraph materials in PR 31026 to ensure speedtree vertex shaders can make use of this keyword.

The reason for requesting this change for the remaining handwritten shaders is to ensure correctness of DOTS instancing in future changes that add indirect draw commands to BatchRendererGroup.  In these future changes, the instance index is packed with a LOD crossfade value in GPU data, so all stages need to be aware of this keyword to ensure correct (and optimal) unpacking.