Fix GPU Hang on Hair Material Pre-Integration
This PR fixes a GPU hang that was happening for a very expensive hair material look-up table computation that happened on renderer initialization for low end hardware. We fix it by generating the look-up table offline and writing it to disk as an asset, then load it with the SRP resource loading mechanisms to get it back into memory and bind it at runtime. 

Some simple infrastructural code was added to help developers in the future pre-bake their render pipeline material look-up tables with this menu item:

![image](https://media.github.cds.internal.unity3d.com/user/1839/files/7c5a3655-6afd-46b1-830d-d12ceff9efbb)

This menu item will automate the transfer of look-up table's created in the new `BuildOffline` interface for `RenderPipelineMaterial` from GPU memory to disk, dumping all of them to a folder in this directory: `HDRPDefaultResources/Generated/`.

It's then the developer's responsibility to take the generated table and load it with HDRP's resource loading system. 

Below is an inspector window of the hair LUT assets that use this new utility.

![image](https://media.github.cds.internal.unity3d.com/user/1839/files/6424dfd5-82cb-4ce3-9264-eac3125983ae)