[HDRP][PathTracer] Refactor path tracing code to be iterative
This refactor removes most of the iterative ray tracing from the path tracing, as specified in GFXRT-100. The only remaining recursion is for subsurface scattering. The goal of making the path tracer iterative is to simplify (and therefore optimize) the closest hit shaders, as well as allow further restructuring of the path tracer into a wavefront path tracer.