[URP] RenderGraph CopyDepth pass injection point location now matches the non-RenderGraph path
The RG and non-RG paths CopyDepth pass location in the frame is different, the RG path one must match the non-RG path for consistency with projects ported from the old renderer.

This PR moves the CopyDepth pass to be executed before the Skybox pass, which is the same behaviour as the non RG path.