[Port] [2023.2] Templates window [+] button now always insert a template except if CT…
Jira: UUM-47255

♻ This is rather a behavior change than a bug fix.
Previously the ➕button used to open the templates window in the last mode (either insert or create).
But when you were clicking on it for the first time in an editor session, it was simply opening the drop down to let you choose which mode you want to open the window in. This could be seen as unexpected since it's the purpose of the dropdown button 🔽next to it.

Now, the ➕button always opens the templates window in `insert` mode, except if you hold the `CTRL` key, in that case the window is opened in `create` mode.

💬 The new behavior is explained in the tooltip.
📘 The documentation has been updated accordingly.