Fixed hash function for PS5.
Backport https://jira.unity3d.com/browse/UUM-38845

The original artefact that caused the test to be disabled on PS5 did not reproduce anymore. Instead a new bug appeared which is related to the computation of noise values on PS5. The PSSL shader compiler was unhappy to divide by float(0xFFFFFFFF), and failed to generate any noise pattern. The workaround is to replace float(0xFFFFFFFF) by float(0x00FFFFFF), which is the maximal continuous integer value that can be represented as a float (mantissa is 24 bits).

Contrary to other backports for this bug, 010-BRG-Simple does not exist on this branch and is thus not an instability.

Reference image for PS5 HDRP Runtime test 009-SG-FullScreenTarget was incorrect and did not have procedural noise working: new reference image fixes this.