Re-enabled 009-SG-FullScreenTarget test scene for PS5
https://jira.unity3d.com/browse/UUM-38845
Re-enabled 009-SG-FullScreenTarget from HDRP_RuntimeTest for PS5.

The original artefact that caused the test to be disabled on PS5 did not reproduce anymore. Instead a new bug appeared which is related to the computation of noise values on PS5. The PSSL shader compiler was unhappy to divide by float(0xFFFFFFFF), and failed to generate any noise pattern. The workaround is to replace float(0xFFFFFFFF) by float(0x00FFFFFF), which is the maximal continuous integer value that can be represented as a float (mantissa is 24 bits).

After 009-SG-FullScreenTarget has been re-enabled, 010-BRG-Simple became an instability on NGGC only and has been disabled (upload of instanced color data in GraphicsBuffer not always fully synced when GPU access it). A separate Jira ticket has been opened to fix the later: https://jira.unity3d.com/browse/PLATGRAPH-2249.