Migrate common render graph resource data
There are a number of places where code cannot be shared because of the addition of ContextContainers. The data common to both the 2D and 3D renderer needs to live in a shared location.

Addresses this card: https://jira.unity3d.com/browse/D2D-5326

Related to this POI: https://jira.unity3d.com/browse/POI-817