[URP][RG]  Recreate pipeline when toggling the global RG setting
Disabling/enabling the setting might cause a lot of issues by caching settings or even textures on materials, because some of the behaviours between RG and nonRG differ. We have issues with bloom texture being cached on the material and a totally different issue but due to same cause of caching Use Native RenderPass setting

To guarantee clean slate, this PR ensures OnValidate call for the pipeline, which destroys and recreates the rendering pipeline to avoid such an issue.