[VFX] Fix memory leak of SharedMaterialData for sleeping VFX systems
While adding an [optimization to skip rendering sleeping VFX systems](https://github.cds.internal.unity3d.com/unity/unity/commit/2a83d4c5e643318f72a69a96cc6c75a81d43ca74), we accidentally added a memory leak for SharedMaterialData. It happens when the base renderer adquires this data and fails to release it, due to a change in the material count that happens inside VFXRenderer::AddAsRenderNode.
The solution was to manually backport a [more recent version of this optimization](https://github.cds.internal.unity3d.com/unity/unity/pull/33367), that landed recently on trunk. In that version, the number of materials rendered is the same, but we manually release shared material data for sleeping systems, replacing it with an empty shader.