[URP] Force shader branching for vrr rendering
Fix VRR performance regression on URP shaders.
This PR improves performance by forcing the branch in the final shader code. This way costly VRR related code is skipped over instead of always being evaluated even on projects that don't make use of the feature.

That being said, even with these changes dynamic branching for VRR introduces a regression on performance which was accepted when choosing the feature in the original PR. It is suggested that VRR be completely excluded from OSX and IOS shaders as those platforms are used only for internal testing to fully mitigate performance loss.