[2023.3][URP] Improving shader stripping on Vulkan in Deferred
This is the second part of the fix for UUM-47782.
In the previous PR (#37827) the issue was resolved but it is also now possible to improve the shader stripping.

Right now for Vulkan, URP always includes the OFF and ON variants for Accurate G-buffer normals due to an issue with one texture format. Now since that format is no longer used, the stripping can have the same check for every Graphics API.