[2022.3][VFX] Fix decal normal issues + filter unused sliders
This PR fixes an inconsistency with the VFX decal normal maps when the particles have a non uniform scale. After this fix, the results are consistent with the non-VFX decals, in HDRP and URP.
Additionally, it removes unused sliders (smoothness, metallic) when they are not used in the URP Decal output.