[URP] [Backport 2023.2] Fixed Color Grading Mode when HDR output is active for the whole stack
When using HDR output on URP, the color grading mode is forced to High Dynamic Range. Using Low Dynamic Range would clamp the tonemapping values and reduce the dynamic range. That's applicable to all cameras in the stack, even if they don't render to the screen directly. Because if one camera clamps the dynamic range but its result gets passed to the last camera that outputs to HDR, then the result is still half-clamped.

To fix this, we don't check whether the current camera is outputting HDR before changing the color grading mode, we check whether the whole camera stack is. The output of a stack is set on the base camera only so no need to check every overlay camera in the stack.