Expose Dynamic Scale Explicit Flag in Render Graph
POI: https://jira.unity3d.com/browse/POI-582 (STP)

This change updates the render graph related code in SRP Core to allow users to control the dynamic scale explicit property of the RenderTexture interface.

When the explicit mode is used, the SRP code gets to decide exactly when the underlying resize event happens on the RenderTexture. This allows dynamically scaled targets to be used with temporal effects since SRPs can now guarantee that their history data won't be invalidated before it's processed.

The ability to preserve history across resolution changes is a pre-requisite for dynamic resolution support with STP.