[VFX Graph] Refactor undo/redo system
This PR reworks the VFX Graph undo redo system
- Fix https://jira.unity3d.com/browse/UUM-29244
- Fix https://jira.unity3d.com/browse/UUM-42481
- Fix other bugs not logged (update of enabled blocks for instance)
- Fix fast path for param changed
- Add fast path (not triggering a recompilation) for node position and collapse state changes
- Improve naming of undo commands (Has the asset name so that it's less confusing with multiviews)

Issue remaining that were investigated but not fixed in this PR:
- Undo stack does not survive domain reload
This is a bit more involved to fix as it is tied to the way our undo redo is hooked into Graphview. To make that work, some changes are required to out Graphview integration. It is orthogonal to the fixes in this PR.
- Some fast path are still missing for everything UI related (sticky notes, group nodes and properties) which means that when undoing/redoing this change, the full graph is serialized and a recompilation is performed even if it shouldn't be necessary. More fast path could be implemented later on.