[VFX] Fix Curve & Gradient edition with instancing
Fix corner case issue introduced with this [change](https://ono.unity3d.com/unity/unity/pull-request/142340/_/graphics/vfx/feature/instancing/share-bake-texture), the update of constant curve & gradient with editor opened was working only when the first instances of every batch were indeed in culling.

It didn't cause issue in runtime because the expression is forced to be updated during first frame.

This solution uses a lazy initialization of `m_MainInstanceIndex`, the first evaluate expression is taking the ownership:
https://github.cds.internal.unity3d.com/unity/unity/blob/42d75480d0aca20a0fa5d3455a2d364e5399ad5f/Modules/VFX/Public/VFXTextureGenerator.cpp#L30
This operation is atomic because evaluate expression can be done in worker thread (@gabriel-delacruz please double check the validity of this approach)

🎁  Bonus, fix two untracked leaks:
- 64 bytes per VFXBatch: https://github.cds.internal.unity3d.com/unity/unity/blob/42d75480d0aca20a0fa5d3455a2d364e5399ad5f/Modules/VFX/Public/VFXBatch.cpp#L112
- 456 bytes per VisualEffectAsset:https://github.cds.internal.unity3d.com/unity/unity/blob/42d75480d0aca20a0fa5d3455a2d364e5399ad5f/Modules/VFX/Public/VFXBatchedEffect.cpp#L47