[HDRP][Yamato]Fix performance test project
HDRP performance test are broken since 27 september, we don't know what have introduce this problem. But this PR fix it.

There was a problem in Forward and deferred asset of performance test causing one variant to not be created at runtime (a variant with screen space shadow)

This PR also update all material and fixing some missing meta.