Add test/scene/ref image logging info to gfx test cases
Users understandably want to see the test/scene/ref image logged to the player/unity log when gfx tests are running. This PR adds that.

https://jira.unity3d.com/secure/RapidBoard.jspa?rapidView=5172&projectKey=GFXGTF&view=detail&selectedIssue=GFXGTF-65

This slack thread captures the discussion around this more. TLDR is graphic test case is an object, not a base test, and as such, doesn't have a nice place to insert common logging like this for now.