[RPW] Fix nullref from CurrentPipelineHelpURLAttribute when no pipeline is active
This PR fixes https://jira.unity3d.com/browse/UUM-53661.

CurrentPipelineHelpURLAttribute is meant to provide an URL that points to either HDRP or URP depending on which is active. It would return null if no pipeline is active, but trunk code assumes that the URL cannot be null. So we return `string.Empty` instead.