Fix Consistency of TAA Frame Index Value
This PR fixes the logic for HDCamera's taaFrameIndexValue to
ensure that:

- It remains set to zero when TAA isn't enabled
- It's reset to zero whenever history is reset (new camera or new cut)

This PR also removes the logic that limited the value to 8 since the
TAA code that calculates jitter already has logic to limit this to the
0-1023 range when computing halton sequence values. That logic was
previously not working properly since it was being artificially limited
to 8 upstream in the camera code.

As of this change, the TAA jitter behavior in HDRP should be predictable
enough to support graphics tests that rely on having a specific jitter value
in order to pass.