[URP][2023.3] Fixing an issue with Panini Projection and disabling four HDR tests on one platform.
### HDR Test Failurres
Recently we fixed an issue where URP RenderGraph jobs were not correctly enabling RenderGraph in the Post Processing jobs. That fix however uncovered that three HDR tests are failing on one platform. For context, [here is a discussion on Slack](https://unity.slack.com/archives/C02LJ5VSV97/p1698328607813429) regarding this.

The tests failing are:
* BluePrimaryIsCorrectColor
* GreenPrimaryIsCorrectColor
* RedPrimaryIsCorrectColor
* YellowIsCorrectColor

Created UUM-55344 for this issue.

### Panini Projection failure
On top of that there was also an issue with how URP enables the PaniniProjection keywords in RenderGraph. It was using a command buffer global but the keyword is set to local so it needed to be changed to use the material for enabling the correct keyword.