[HDRP] Framesettings overlay
Goal is to provide feedback when a setting depending on camera position/camera settings is disabled for an active effect

It handles framesettings disabled from the global settings, from the rendering debugger, from the camera inspector, or because there is a dependency on another framesetting.

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/4819e5e1-b2d2-4c5e-aff6-91680ee31518)

Additionally, for effects depending on volume settings (that are camera dependent), we also inform them that the effect is not enabled.

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/819a4179-9251-4090-b887-761f2c7c9d0f)

Note that there are cases where you actually want to have those effects/framesettings disabled so it's not an error.
And in those case, there's not much point in openning the inspector for the effects, so it shouldn't cause any workflow issue for users.

This is partially implementing the design made here https://www.figma.com/file/LDBHk0xvnf7rsBbLIZwr0W/Frame-Settings-HDRP-global?type=design&node-id=165-16871&mode=design&t=od2hhKVh4LMIv1Rf-0