[SRP][TestsOnly] Remove orphan meta file
Found by @motiejus-viltrakis on Slack:

> Motiejus Viltrakis
  [1 hour ago](https://unity.slack.com/archives/C03JBLHEMNV/p1699353790180679)
This meta file doesnt have its "partner". Can it be deleted? Its resulting in a warning on trunk
https://github.cds.internal.unity3d.com/unity/unity/blob/trunk/Packages/com.unity.render-pipelines.core/Tests/Editor/TestComponents.meta

This was an file accidentally added in PR https://github.cds.internal.unity3d.com/unity/unity/pull/38997. The meta file corresponds to a folder that doesn't exist, so I'm removing the file.