[APV] Improve sampling noise
Sampling noise was always applied along vector (1,1,1), which didn't make much sense. It is now along the direction going towards camera, which reduces leaking due to sampling noise.

For example in a scenario like that where there is a wall or a floor, and very different lighting intensities on both sides, the noise could make APV be sampled from both sides alternatively and produce a lot of leaking
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/06c840d0-5fcf-407a-80ea-d944918ca0bd)



| Before | Now |
| -- | -- |
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/ef3e3ae5-eeb6-45e6-b9ed-a1cceef42c29) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/30c57822-8657-4a05-8cd2-2409eb398bc7) |

Additionally:

- fixed a tooltip mentionning HDRP in URP
- Made handles bigger for volumes
- Made handle color more visible
- Reenable hardware backend for virtual offset
- Removed advanced label on baking sets
- Added sky occlusion weight in volume system