No Streaming Asset path for APV
https://jira.unity3d.com/browse/HDRP-2997

The goal of this PR is to add an optional code path for APV in which the data is not stored as StreamingAssets in a standalone build. This will make the APV data compatible with Addressables and Asset Bundles at the cost of Disk Streaming.
One drawback of this option is that it's now impossible to partially load APV data in memory since they are regular assets. This means that in multiple scene setups, the memory footprint is that of the full baking set instead of just the loaded scenes.
This new optional code path can be enabled in the Graphics Global Settings in the new Probe Volume tab. This should also work for URP by default.

To achieve that, several refactors were required:
- Removed all serialized data from ProbeVolumeSceneData. The data is now stored in their respective Baking Sets in the BakeData class.
- Added a new ProbeVolumeGlobalSettings class which contains the setting for the new code path.
- Moved all APV resources (shaders, meshes and textures) to Core through the use of the IRenderPipelineResources interface.
- To implement proper stripping of resources, added a new IIncludeAdditionalRPAssetsOnPlayerBuild  interface to handle the special case of HDRP when including SRP assets for the build (scene dependencies and labelled assets). This could be used by URP in the future.
- Refactored the Init/Deinit of APV to make it cleaner (and fixes assets in multiple scene setups)
- In the new code path, now we reference the actual text asset in the ProbeVolumeStreamableAsset in a build step in order to properly bring the dependency into the player (which means that the data will always be fully loaded when the baking set is loaded).
- Split all Editor specific code in a separate Editor only file for the ProbeVolumeBakingSet class.
- Data from the previous version should be migrated properly (mostly data like sceneBounds that were stored in the global ProbeVolumeSceneData should be properly copied into the relevant baking sets).
- Removed some very old migration code.