[Port] [2023.2] Fix ShaderGraph dirty when opened for the first time and requiring a checkout of VC
Fix ShaderGraph dirty when first opened.

Note that I needed to change the hashcode computation of some properties to make sure the hash is always 0 by default. This ensures that a new ShaderGraph will have the exact same hash as the serialized one when the graph is created on the disk.