Remove extra CommandBuffer
Currently the NRP compiler has a layer of command buffer on top of everything else.
This bakes render graph per frame data in a command buffer which makes it impossible to actually reuse the result.
Since this layer is only added overhead (what it does can be done in Pre/Post pass execute functions), we need to remove it before we can do the caching.