Light transport/cleanup light transport package folders
https://jira.unity3d.com/browse/LIGHT-1752

This PR moves around some code in the light transport package. Code unrelated to raytracing (intersections) has been moved out of the UnifiedRayTracing folder, and into a new Sampling folder. Most of the code has just been moved directly without any changes, though I had to make incisions in a few places:
- I have split Common.hlsl into 2 files - 1 that contains common code for raytracing, and the other for sampling.
- I've changed the namespace of the moved C# code, and moved a function for binding sampling textures to a different class, in that new namespace.

This change is in preparation for landing the new baker we are working on in `light-transport/unified-dev`. I'm essentially dealing with upcoming conflicts early.