[HDRP] Merge the implementation of rect and line lights
POI: https://jira.unity3d.com/browse/POI-816 - SRP production ready

The goal is to essentially halve the size of both the written and the executable code (good) at the cost of a few static branches (bad). I expect the performance to be largely the same. The maintenance burden is reduced; the same applied to the compile times and the shader binary size.

As a part of this change, I introduced a "pillow" windowing function that is used for range attenuation of both rectangular and line lights. The principal advantage is that it behaves the same way as the windowing function of point lights, so now the attenuation behavior is consistent across all supported local light types.