[HDRP][PathTracer] Re-ordering throughput multiplications to allow better clamping
Due to the recent refactoring of the path tracer code, the clamping applied to the different contributions became worse. In this PR, some of the necessary multiplications are re-ordered to allow clamping light values *with* the BRDF value before the light intersection included. This leads to more effective clamping of fireflies and reduces the bias introduced, as is clear in the below pictures. 

Image without any clamping applied: 
![final_reference_without_clamping](https://media.github.cds.internal.unity3d.com/user/6813/files/f12899b3-6703-4a33-a839-391eacdfe2f3)

Image with current clamping:
![screenshot_ref_example](https://media.github.cds.internal.unity3d.com/user/6813/files/b32b5fce-0c5d-4327-b3e0-1b80104d6551)

Improved clamping with this PR: 
![screenshot_new_example](https://media.github.cds.internal.unity3d.com/user/6813/files/eb0dc5de-7406-4072-a990-e91660b51669)