GPU driven Tier 1 occlusion culling
- EPIC: [GFXRT-46](https://jira.unity3d.com/browse/GFXRT-46)]
- POI: [GPU-Driven Architecture - Tier 1](https://jira.unity3d.com/browse/POI-926)
- [GPU Driven Tier 1 TDD](https://docs.google.com/document/d/13Hm-Pc0Xml8-mB80FHRnZy5uDHkGuTqfhpqQLOO3TGM/edit?usp=sharing)

This PR adds GPU occlusion culling support to the GPU Resident Drawer.  This allows for GPU-side culling of instances that do not contribute to the final image.  This can help to improve GPU performance in scenes with heavy overdraw, and does not require any additional authoring to try out.

This functionality is exposed as an API in the core SRP package in order to integrate with custom SRPs, but this PR also contains:
- Integration with URP (requires Forward+ and render graph enabled)
- Integration with HDRP
- Rendering debugger integration (for both URP and HDRP):
  - Statistics about the number of instances that pass/fail occlusion culling
  - Visualisation of the occluder mip pyramid
  - Heatmap visualisation of culling due to the occlusion test

Screenshot of rendering debugger integration within URP, showing the occlusion test heatmap and tables showing the number of culled instances (around half of the instances in the frustum are culled):

![image](https://media.github.cds.internal.unity3d.com/user/4787/files/93eabae1-ea10-48fc-af02-ce2357c6fe1f)

In both URP and HDRP, this PR allows the user to enable occlusion culling via checkbox on the SRP asset:

![image](https://media.github.cds.internal.unity3d.com/user/4787/files/55ec9321-8b26-48f6-8670-c17275faa8de)